% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_simulation.R
\name{parallel_simulation}
\alias{parallel_simulation}
\title{Perform Parallel Simulations for Bayesian Platform Data-augmentation}
\usage{
parallel_simulation(num_simulations = 4, num_cores = 4)
}
\arguments{
\item{num_simulations}{An integer specifying the number of simulations to run.
Default is 4.}

\item{num_cores}{An integer specifying the number of CPU cores to use for parallel processing.
Default is 4.}
}
\value{
This function does not return a value directly. It saves results to CSV files
in the working directory.
}
\description{
This function performs multiple parallel simulations of A Bayesian phase I/II platform design with data augmentation
accounting for delayed outcomes (BPDD).
The function executes a series of simulations using the main simulation function
\code{main_BPDD} in parallel across multiple cores, generating various statistics and
saving results to CSV files.
}
